Add formatter for javascriptreact filetype
Something has decided that JSX files should have a specific filetype in
vim, ie not just `javascript`. Which isn't unreasonable, as it's not
conforming to ECMAScript standards any more.

However this means that `.jsx` won't get formatted using `prettier` (via
neoformat).

Therefore we're adding a clone of the Javascript config to resolve this.