Only activate Chrome if it's off screen
I use `<Leader>ww` to rebuild/reload the file that I'm working on. For web
stuff this usually means reloading the browser. Depending on how large a
screen I have I may or may not have the terminal taking up all my
screen.

If it is on screen I don't want to `Activate` the window as that will
take the focus out of vim. However if it's running in the background
then I do, as that will switch to the workspace that it is running in. I
can then ⌘+Tab back, which feels like less of an imposition in that
case, for some reason.